Move API route handlers into api.go
To separate Model definitions away from the transport interface, move
the route handler functions into their own file.
